RONDB-282: Parallel copy fragment process Copy fragment process have been limited to one fragment copy per LDM thread previously. The copy fragment is also limited by the parallelism such that at most 6000 words are allowed to be outstanding (a row counts for 56 words plus the row size in words).
This means that in particular initial node restart is fairly slow. Thus we
should improve the parallelism while at the same time maintaining protection
against overload of the live node which could be very busy serving readers
and writers of the databases.

Actually this parallelism is already implemented in DBDIH. This means that
we can already send up to 64 parallel copy fragment requests to a node.
However DBLQH imposes a limit of one at a time and queues the other requests.

This patch ensures that we can run up to 8 parallel copy fragment processes
per LDM thread. We will get data from THRMAN about CPU load due to this,
we will use this both to limit the number of concurrent copy fragment
processes and also to limit the parallelism per copy fragment process.

One problem when synchronising a node and there is a lot of disk columns,
is that we might get overload on the UNDO log. This would cause failure
of the node restart if not avoided. To avoid this we need to be able to
halt the copy fragment process and later resume it again.

This functionality was already implemented. However the implementation
isn't very likely to work. The new implementation is very simple. It
sends a HALT_COPY_FRAG_REQ to the live node(s) performing the copying.
It doesn't send anything back to the starting node. It is the responsibility
of the live node to halt in a proper manner. Similarly the starting node
will send RESUME_COPY_FRAG_REQ when it is ok to copy again. Currently
the halt will halt all copy fragment processes even if they don't have
any disk columns. It is likely that the disk columns tables are anyways
the bottleneck for the node restart, so special handling wouldn't make
any major difference.

To handle up to 8 parallel copy fragment processes per LDM thread requires
that we have reserved 8 operation records, there is a scan record in
DBTUP, there is a scan lock record in DBTUP, there is a stored procedure
record in DBTUP, there is a scan record in DBLQH and there is a scan
lock record in DBACC, all of them requires 8 reserved records such that
we don't risk that the copy fragment processes runs out of memory.

A few tweaks were required to send COPY_FRAGREQ for NON_TRANSACCTIONAL
use. This is part of restore fragment and is mainly used in initial
node restart. Some tweaks were required here to ensure that we could
start more in parallel.

Both TRANSACTIONAL and NON_TRANSACTIONAL COPY_FRAGREQ required limitations
when running against an older node starting up.

A very important part of the work is to also adapt the speed of node
recovery based on the CPU usage in the live node. We need to slow down
when the CPU gets heavily used.

The speed of node recovery was previously limited to having 24 kBytes
outstanding where 0 bytes in a row was counted as 224 bytes. This
means effectively less than 50 rows at a time per thread was outstanding.
We increased this substantially to now go up to 192 kBytes instead.
Thus around 400 small rows can be outstanding. This should provide a
good balance between latency of operations towards the live node and
the speed of node recovery.

The CPUs have become much faster and on some fast CPUs adding microseconds
wasn't good enough, we had to raise the level to handle nanoseconds.
This had an impact on our measurements of CPU load which is a very
important part of this patch for adaptive speed of node recovery.

As part of this patch we also disable ACC scans (previously done in
22.10 branch).