RONDB-364: Add service-name parameter to ndb_mgmd and ndbmtd
This parameter adds a --service-name parameter to ndb_mgmd and ndbmtd.
E.g. --service-name=ndbmtd sets the file name of the pid file to ndbmtd.pid
and node log to ndbmtd_out.log and similarly for trace files and other log
files and the error file. Also sets the directory name of the NDB
file system to service_name_fs.