RONDB-371: Zero files in optimised manner at initialisation
When initialising REDO log files at startup we want to ensure that the
entire file size is preallocated and that the file is properly initialised.

The same is true for UNDO log files in LGMAN and tablespace files.

To speed up this we use special XFS methods and POSIX methods that apply
to the file systems we recommend users to use. This means that we can
preallocate file space without actually writing it and we can even ensure
that reads of this area returns pages filled with zeroes.

We still need to write the initial parts of the file and for the REDO log
we have to write each start page of each MByte.