RONDB-398: Decreased delay to check receive to 1 ms
Updated max wait time for checking receive to 1 ms. Ensuring always
the lowest latency possible. Most of the time it should not matter,
but receive thread is the most important thread to wakeup quickly.
So some extra safeguard around it is a good idea.

Also updated m_max_send_wakeups, not expected to need any updates
to it, it is mostly to protect from cases where an extreme amount
of messages are queued up, to ensure that we send response in a
timely manner also in those cases.