RONDB-440: Fixed a hole in fragment protection for multi range scans
When performing a multi range scan we setup the second range scan using the method
that also acquire the fragment mutex if required. However immediately after that
we released the fragment lock and returned, after the return we continued scanning
the new range, but now without the fragment lock since it was released.

This can cause problems only in Query thread scans and only with range scans scanning
more than 1 range and of course executing at a time when someone is changing the
ordered index.