RONDB-479: Double mutex lock led to watchdog failure
When an LCP needs to allocate a page to handle a DELETE it could run out of
Undo pages, to avoid that we release the page before we allocate the copy
page, this needs to be done under mutex protection. The handling of this was
buggy such that the mutex was taken twice leading to a watchdog crash.