RONDB-155: Avoid using non-readable nodes for read
While adding the ability to read with locks on query threads it became
possible to read from nodes that are currently building their data in
the fragment. We have to ensure that for reads and scans that we only
return such nodes in DIGETNODESREQ that have all data and thus is
readable. This means that we disallow nodes added in the STORED phase
of UPDATE_FRAG_STATEREQ signalling and that are completed in the
COMMIT_STORED phase. Thus when a replica reaches the COMMIT_STORED
phase it becomes readable (although not necessarily durable yet).

This also fixes an issue with Location Domain Id's where it was also
possible to issue reads against nodes that are not yet readable.