RONDB-486: Map node group to start from 0 and be consecutive
It is allowed to set Nodegroup id on a node. However for DBDIH to work
these ids must start at 0 and be consecutive. Otherwise things will fall
apart in lots of places.

To make it easier for the user we add a function that maps from node group
to a node group that starts at 0 and continues consecutively. Thus for
example if the user sets Nodegroup to 1 and 2 it will be mapped to 0 and 1.