RONDB-531: Give responsibility to create RR groups to NdbHW portability layer
NdbHW.cpp has created the L3 cache groups used to create the CPU map that is
used in assigning block threads to CPUs. Thus it is much easier to allow
for NdbHW.cpp to also set the m_rr_groups used in block threads to figure
out which RR group a certain signal should be sent to.