RONDB-531: New send handling
With this patch the handling of send threads is changed.
Every time we send we set a timer that disallows any
send on that transporter to happen again until the
timer has expired.

In addition when we insert a transporter into the ready
transporter list we extend this timer by up to 30
microseconds to ensure that we will wait with sending
until all threads have handled the received signals.

The timer is set dependent on the current load on the
send threads and the send thread assistance from block
threads.

We can increase the timer up to 250 microseconds at
very high send loads. At overload we increased the
latency to 500 microseconds.

We have also made it possible to set the maximum send
delay through a DUMP command to accomodate for some
experimentation on what settings of the timer that
works well.