RONDB-531: Introduced MaxMicrosAwake configuration variable
To avoid spinning for extended period of times without going to sleep
we added a new configuration parameter that specifies the max number
of microseconds we will be allowed to be awake to start spinning.
That means that if we have no work to do, and we have already been
awake for an extended period of time, we will go to sleep. This should
ensure that the OS gets a higher possibility to schedule its
activities when the data node isn't busy.