RONDB-531: Elaborated on send delays
Experiments have shown that at very high loads send delays can provide significant
throughput improvements. It is however very difficult to understand when to increase
the send latency since increasing it will decrease the send load. Additionally there
is no direct correlation with CPU usage.

In this patch we add two extended delay parameters. The first sets the extension
of the send delay that we provide when we alert the send threads. The second
sets a limit to the number of times that we are allowed to extend the send delay.