RONDB-531: Ensure load is evenly spread among query workers
Previously we spread the load in a manner that preferred the efficiency of
using the LDM owner. This was good for efficiency but meant that the
LDM owner had much more work to do and this led to unbalanced load on
the CPUs. This in turn led to bottlenecks and thus not using all CPUs.

Now every second scheduled query will use any query worker and only
half will be targeting the LDM owner and its shared LDM instance.

This should be a good compromise between efficiency and fairness.

Also fixed a minor issue in that if the distribution array shrunk
such that the current index is outside the boundary we will start
from 0 again. This will only affect a few LQHKEYREQ signals and
thus not really be any issue, but still good to be correct.