RONDB-557: Improved send delay handling
In this patch we ensure that at high load we increase latency by
waiting with sends to ensure that we gather enough messages to
avoid swamping the CPU with interrupt handling. In some benchmarks
we could see that each CPU had to handle either a transmit or a
receive interrupt every 30 microseconds.

The send delay is controlled by 3 parameters at high load:
1: The send delay, this is adaptable and computed by thrman.cpp
based on send load and CPU load. It is a timer that is reset each
time we send on a transporter. When the timer is set it isn't
allowed to send on the transporter until the timer has expired.

2. The extend delay. Every time a thread calls alert_send_thread
on a transporter we extend the send delay by the extend delay.

3. The max number of times we can extend the delay.

Optimal numbers we found by experiments were
100 microseconds send delay at medium load and 125 microseconds at
high load.
8 microseconds in extend delay that is allowed to be extended up to
8 times.

If the time has expired we will not set the send delay again until we
have sent on the transporter.

All these three parameters are settable through a DUMP command to
make the parameters testable. These settings are protected against
setting very bad values.

RONDB-557: Improved send delay handling