RONDB-563: Smoothing LCP load
Removed 25% extra LCP speed to accomodate variability. This is only added when
we come to at least low or medium usage of the REDO log.

Removed in a similar way extra LCP speed to accomodate increase of lag. This is
also only added when low or medium usage of the REDO log.

Thus the idea is that we run LCPs at a much lower speed when the REDO log isn't
at any risk of running out.