RONDB-567: Bugs in List management in SUMA
As part of poolification of ReplicationMemory the free list in SUMA
of pages for events was changed from single list to double list.
However the handling of previous pointers was not done in a proper
way. Also an old bug when failing to allocate pages from global
memory was fixed.