RONDB-556: Fix of adjust_weights
adjust_weights affected the stored weights, this is not a good idea
since it could lead to that some threads never get smaller than
others once they have received a high load.

As an example if a thread goes at high load where it gets 8 in weight
and the other are 2, in this case the high load gets to 16 and the
other one goes to 4, this means that the high load thread always
gets a new increase once it gets back to 8 again.