RONDB-556: Handle Round Robin groups independent of each other in query distribution
It is important to not make to quick movements of weights to adapt to the environment.
To handle this we only move at most one step per time unit (50 ms) to adjust towards
the middle. This prevents heavy weights to stay heavy weights even when overloaded.

Threads from different round robin groups are independent in scheduler, so need to be
independent in the query distribution scheduler as well.