RONDB-556: Ensure Round Robin groups are balanced
When using automatic thread configuration based on HW configuration we
will assign threads to CPUs based on where the L3 caches are accessible.

Each LDM thread is assigned to a Round Robin group and each block thread
is assigned to a Round Robin group. Query workers can only assist LDM
threads in the same Round Robin group.

This means that it is essential that the LDM threads are spread into
the Round Robin groups in an even fashion. Since we assign CPUs in
pairs (to handle CPU cores with hyperthreading when they are used)
this means that we want to have an even number of LDM threads and
we do some mathematical calculations to ensure the LDM threads are
as balanced as possible in the Round Robin groups.

In this manner we try to achieve a balanced load on all threads
by ensuring that we have a balanced use of each Round Robin group
and that the CPU resources for each Round Robin group is also
evenly balanced.