RONDB-581: Set more reasonable memory settings
Ensure that MaxNoOfConcurrentOperations setting means that we have memory
to handle a large transaction with that many operations.

Decreased amount of memory that is reserved for ReplicationMemory,
SchemaMemory, BackupSchemaMemory, JobBuffer and Send Buffer. Added
memory from the previously reserved memory to shared global memory
to ensure that there is more shared global memory to use when a
resource runs out of reserved memory.

Move TransactionMemory from ultra high prio to high prio since
having TransactionMemory run low doesn't risk crashing the node
as it does for JobBuffer, SendBuffer and ReplicationMemory.