RONDB-583: Fixed a problem with memory management in SUMA
In communicating between TUP and SUMA we use special trigger pages.

Those pages are allocated from TUP and released in SUMA. We could allocate 1 or 2 pages
in TUP. The problem is when we allocate 2 pages and only 1 remains in the chunk.
In this case we will insert the page in the full list. When we release a page in SUMA we
will move it from full to free only when there is 1 page in the chunk. However if there was
one page remaining and one was free’d, then there will be 2 free pages in the chunk.
Similarly if the chunk was empty and we release 2 pages at once, in this case as well we will
not move it from full to free. Finally we will release the last page and in this case we will
release the whole chunk if there is one chunk in the free list. Thus if we release the chunk
we remove it from the free list, but it is in the full list, thus we will corrupt lists,
if we don’t release we will leak the entire chunk since there is no other way of getting
removed from the full list.

Fixed by always moving from full to free when releasing a page. Also increased chunk size
to align with size used by RonDB mallocator.