RONDB-584: Added method to verify page is not in the set of free pages in global memory manager
To ensure that we are not working on the free pages as if they were allocated we added a call to the global memory manager to verify that a page is allocated, this checks that the bit in the bitmap of free pages is not set.

Later we can add calls to perform this check, in this patch we add it for the SUMA trigger pages.