RONDB-584: Make it possible to verify pages are allocated when used
Previously the free bitmap of pages in global memory manager only
set the first and last bit of page ranges. This meant that the
bitmap could not be used to quickly verify if a page has been
allocated or not when it is used.

We changed this such that every bit in the bitmap now indicates
if the respective page is free or allocated. Thus the new method
verify_page_allocated can be used to check if a page is allocated
or not. This will make it a lot easier to find cases where we
use a page after we released the page. These types of bugs are
otherwise extremely hard to find.

To handle this efficiently an implmentation of clearRange of a
bitmap was added together with a unit test for it. The test was
executed and passed before committing the patch.