RONDB-585: SUMA can free pages still in use + a memory leak
When allocating a new chunk of pages it is put into the
free list in SUMA, but it is also kept in the free chunk
list. This means that if we allocate half of the pages in
the chunk and then the first free_page call can at times
call release_chunk and release the entire page chunk since
the chunk is in the free chunk list, but some of its pages
is still in the free list.

One other thing fixed in this patch is to avoid calling
remove from free chunk list when last page is allocated
from chunk. The reason for this is that it isn't in this
list and calling the remove function will effectively make
the list empty, thus a memory leak.