RONDB-136: Support more than 20320 table objects
The default setting remains set to support 20320 table objects. However by
increasing MaxNoOfTables, MaxNoOfOrderedIndexes and MaxNoOfUniqueHashIndexes
one can now make it possible to handle more than 20320 table objects.

In addition to changing a new configuration variable is introduced
MaxNumSchemaObjects, this is by default 20320. This sets the maximum
number of table/schema objects that the schema file can contain.

It is necessary to set the MaxNumSchemaObjects to a number larger
than 20320 to enable handling of more than 20320 table objects.
This number must be higher than the sum of the above configuration
variables for number of index and table objects.

This change introduced a number of challenges.

One such challenge is that sending the signal SCHEMA_INFO is a challenge
since it could send 50 MBytes (with 200k table objects) in a single signal.
This required some changes to the method execCONTINUE_FRAGMENTED method
in SimulatedBlock.cpp. This loop sends 240 bytes at a time. This execution
will however continue looping until the full signal have been sent. To
ensure that we don't overload the send buffer we therefore introduced a
check of send buffer level in this method. If send buffer level is too
high we stall sending of that particular fragmented signal.

To support this check we introduced new methods in TransporterRegistry to
get send buffer size and to find the transporter id we're sending to
(this can be a bit complicated due to the introduction of multi
transporters for communication between data nodes in the same node group).

Removed an unused parameter of get_send_transporter method to simplify
things.

Updated the method calculate_send_buffer_level to both be affected by
overall send buffer memory usage, percentage of send buffer in transporter
used and also the actual size of the send buffer queue.

Removed some methods no longer used in Dbdict to avoid concerns about
handling schema files.

The second issue is that during a schema transaction we write the full
schema file which could be 50 MByte, this can easily become a performance
bottleneck. To avoid this we track first and last page to write in the
schema file. This might require more work to optimise it even further.

There was also a need to make a few changes in PGMAN to handle the ordered
lists for handling checkpoints of disk column data.

Added a number of checks to avoid issues with misconfiguration.

Added a bit more weight to the NodeReceiverGroup object to make it
easier to check send buffer level.