RONDB-641: Fixed a number of problems with move of disk row
Normally a Commit of a disk row only involves one page. This case
means that normally there is a single UNDO log record and the
page that diskPagePtr refers to is this single page involved in
the commit of the row operation.

With variable sized disk rows we have a possibility that the row
must move to a new page when it is growing to a larger size. In
this case we need to both log the UNDO of the removal of the old
row and the UNDO of the allocation of the row.

In this case commit sends in diskPagePtr referring to the new
row page. However this was mixed up in the commit handling and
we instead used the old row page where the new row page should
have been used and vice versa.

Also when freeing log space we used the wrong page to figure out
whether to return log space.

A final complexity came from the fact that the UNDO of the allocation
of a new row wasn't part of m_undo_buffer_space, this was accounted
for in one place, but not in another place.

This patch also adds a bit more debugging support required to understand
what was going on in the commit of disk row that used DISK_REORG.
Also some comments with faulty comments or comments that weren't
understandable were removed.