RONDB-641: Fix issues related to multi operations combined with DISK_REORG
When calling load_diskpage we haven't yet called prepareActiveOperation.
This means that we will never have set the bit field
m_load_extra_diskpage_on_commit. This bit exists on the previous operation
record, so it must be fetched from there in load_diskpage if there is a
previous operation.

Second in load_extra_diskpage we need to get the pointer to the copy
tuple. This is however not present at this early stage in the new
operation record. Thus this must also be fetched from the previous
operation record.

Finally in handle_size_change_after_update we need to use the
extra disk page when DISK_REORG is already set. This was the case,
but we calculated some variables using the wrong page, we used the
old page instead of the new disk page. Thus these variables needed
to be recalculated when discovering that the DISK_REORG bit was set.

Also added some more jam's in debug and in production.

Forgot to remove declaration of useless variable causing crash

Fixed printout in debug mode that can cause crashes