RONDB-687: Increase timeout before crash
In situations with VMs that have CPUs shared with other VMs we can
suffer from having access to CPUs that are far between and very
short lived when getting it. This can persist for a while, we
increase the timeout check of the scanning for timeout to be in
the order of 10 seconds instead of 1 second. The 1 second has
been seen to cause issues and thus this should resolve this.