RONDB-503: Multi-database transactions and drop database quota improvements
A transaction could use several databases with different quotas, but it can
only be using quotas from one database. The operations in a transaction has to
be executed in the order they arrive and this means that we have to maintain
both order within the transaction as well as in the database.

Our choice is that the first operation that is connected to a database that uses
quotas will use this database for quotas.

Since this means that the transaction has a reference to the database object,
it has impact on the drop of database quotas. To handle this we have to
use reference counting on the transaction object, thus every time a transaction
is attached to a database with quotas, it will increment the reference count
on the database and similarly it has to be decreased when the transaction is
finished.

When the database is dropped, a prepare phase ensures that all transactions
have no more references to the database and also that the database quota
queue is empty and that there are no outstanding queries in the job buffer
waiting to be executed.