RONDB-693: Prepared API Key cache for higher concurrency
Before this patch the key cache map was hardcoded to be one instance
and similarly for the mutex protecting it.

With this patch it becomes possible to change the number of API key
caches to any number on the form 2^n, it was tested with 1 and 4.

Since we in the patch set it to 1, the patch has no functional
change compared to the previous patch, but it makes it very easy
to change if benchmarks shows that the API key cache is a bottleneck.