RONDB-842: Fixing asserts
When sending TRANSID_AI it is fully possible for a node to go down
at any point in time, so an assert of node being up isn't useful.

When a node crash it is possible that the master records the crash
GCI as e.g. 896 while the node has already recorded 897 as completed.
This is discovered while preparing START_FRAGREQ. We will send GCI
897 to the starting node as it is ok with restoring 897 according to
its own data.