RONDB-842: Solving order of UNDO log files and added debugging
Previously all UNDO logs was initialised with LSN = 0 in page 1.
This meant that the order of the UNDO logs is essentially the
order they got at creation of the file objects. This order
is based on the id of the UNDO log file which is not necessarily
constantly increasing. At creation of the UNDO log the first UNDO
log file is choosen as HEAD. However at restart the last UNDO log
file is choosen as HEAD. If the first UNDO log file is filled
most of the problem disappears since the restart will use sorting
and this could salvage the situation.

The remedy of this problem is to ensure that the very first UNDO
log file is initialised with LSN = 1, This means that the sorting
is sure to pick this UNDO log file again as the current UNDO log
file. If the first UNDO log has been filled then the new UNDO log
file can be choosen anyone, but once it is choosen it will write
a higher LSN number into page 1 and thus ensure that it is picked
for the restart as the current UNDO log.

A side effect of this is that we need the first LSN to be 1 and
not 0, so it affects the initial values of LSN.

To ensure higher probability of avoiding this problem in an upgrade
situation we will ensure that if all files are 0 in page 1 the
first UNDO log file will be choosen as the current one.