RONDB-842: Fixed check for UNDO_END with updated LSN numbers
Fixing find of UNDO_END
-----------------------

Previously the last LSN was 0, thus when m_last_read_lsn == 1
it meant that we reached the end of the UNDO log and we return
an UNDO_END log record. Now we should check m_last_read_lsn == 2
instead.

This makes it possible to miss the last UNDO log record in upgrade
cases. To avoid this issue we write in the m_last_lsn in the page
header of page 1. This will be 1 in new code and 0 in old code.
We use the fact that this page information was 0 in old code and
call it m_last_lsn in new and set this to 1. This information from
binary search is used to set m_last_lsn in log group record. The
variable is initialised to 1 and will stay 1 unless an old UNDO
log record is read in which case it will change to be 0 instead.