RONDB-855: Makre sure that ACTIVATE and DEACTIVATE can be executed in a idempotent manner
ACTIVATE can be executed even if the current MGM server thinks it is already active.
The MGM server is not necessarily aligned with the state in the data nodes and API
nodes and thus the action need to be executed even if the state is already active.

DEACTIVATE call will lead to the node being stopped if already started. However a
stopped node can be deactivated multiple times for the same reason.

Set HOSTNAME can only be executed with nodes that are deactivated. A deactivated node
cannot be in the NOT_STARTED state since this state has set up connections preparatory
to restarting the node. The node has to be fully disconnected when deactivated and thus
it is not currently possible to make an MTR test case testing DEACTIVATE calls on data
nodes. These test have been performed manually.