RONDB-860: Use one transaction per operation for RDRS batch requests
When issuing a batch request, it should be treated as a set of pk-read requests
in RDRS. This means that we can use one transaction per operation.

Using one transaction per operation means that we can send the operation to the
nodes where the primary replica resides directly which will cause load balancing
and parallel reads to start already from the NDB API. With a single transaction
all requests have to go through one node and one TC thread.

Here we scale on number of nodes and on number of TC threads per node.

The overhead of a transaction is almost neglible, so no issue on that front.