RONDB-862: Fixed handling of errno
In the case of strtol and strtoll errno is not set before returning
successfully. This means that an old errno value can look like a failed
call to strtol(l). To avoid this a lot of setting errno = 0 was introduced,
in particular at the start of most threads and programs.

In addition some bug fixes around handling return from strtol(l) was
done.