Fix build scripts to not build after failed config
CMake can fail and unfortunately still produce (partial) makefiles. We
don't want to pretend everything is fine and build using such
makefiles. This patch adds `set -e` to the build scripts so that a
failure in cmake causes the script to exit with an error code and not
run make.

Also:
- Use bash
- Fix formatting