RONDB-905: Hint TC placement for REST API server
Previously no hinting of TC placement was done in REST API Server.
The REST API server uses NdbRecord variant of NDB API and there is
a simple way of using NdbRecord and pointer to the row for
startTransaction call. However one needed to move the setting of
the row from setup_read_operation to a new method setup_primary_keys.
Renamed setup_transaction and setup_read_operation to setup_transactions
and setup_read_operations. Fixed problem introduced by previous patch
when setting BatchMaxSize, the change needed to be in init_defaults.go
as well for MTR tests to pass.

The special case where all operations in the batch were invalid was handled
in a few places.