RONDB-899: Improve scaling of large data nodes
As nodes goes larger the receive threads have to split sending to many more tc
threads. This includes a lot of wakeups and synchronization and mutex handling.

In this patch we provide a configuration variable that decides whether we should
only distribute within the Round Robin Group of the receive thread. This has
two advantages, first a receive thread will no longer need to send signals to
a great number of tc threads, instead it will focus on sending to 2-4 tc
threads normally. We also introduce a variable that sets the optimal
round robin group size. Both of these new configuration variables enables
us to experiment in benchmarks the efficiency of this approach.

The second benefit comes from the fact that the Round Robin Group shares
normally at least a L3 CPU cache. This means that communication to the
selected tc threads should be more efficient as well.

There is likely still some scaling issues remaining since there is a
large number of LDM threads still for the tc threads to communicate
with.