Postfix for commit 749558: More restrictive condition for terminating mysqld
The previous patch was intended to terminate mysqld after a timeout when
encountering errno 1101. However, errno 1101 is also used for other
errors related to slot ID allocation on the management node.
What we really need here is to terminate mysqld only for the specific
case of "No free node id found for ...". This patch filters out the
other errors that share the same error number.